+++ site/pages/platforms/next.mdx
@@ -219,23 +219,23 @@ export const GET = handle(app)
 export const POST = handle(app) 
 ```
 
-### Bonus: Page co-location
+### Bonus: Page & Frame Co-location
 
 If a user shares a link to your website, we may want to render the frame in the page route
-itself, instead of forcing him to share a link to the frog api route.
+itself, instead of forcing them to share a link to the Frog API route.
 
-In the example below, when a user shares the `/` (= `/page.tsx` in your Next.js App)
-path of the website on Warpcast, they will be able to see the frame.
+In the example below, when a user shares the `/` (= `/page.tsx` in your Next.js App) 
+path of the website on Warpcast, they will be able to see the frame. 
+This leverages the [`generateMetadata`](https://nextjs.org/docs/app/api-reference/functions/generate-metadata#generatemetadata-function) function built-in to Next.js
 
 ```tsx twoslash [app/page.tsx]
 // @noErrors
-import { extractFrameMetaTags } from 'frog'
-import { type Metadata } from 'next'
-// The base URL of your app
-import { APP_URL } from './app-url'
+import { getFrameMetadata } from 'frog/next'
+import type { Metadata } from 'next'
 
 export async function generateMetadata(): Promise<Metadata> {
-  const frameTags = await extractFrameMetaTags(`${APP_URL}/api`)
+  const url = process.env.VERCEL_URL || 'http://localhost:3000'
+  const frameTags = await getFrameMetadata(`${url}/api`)
   return {
     other: frameTags,
   }
