+++ src/next/package.json
@@ -1,0 +1,5 @@
+{
+  "type": "module",
+  "types": "../_lib/next/index.d.ts",
+  "module": "../_lib/next/index.js"
+}
