+++ src/package.json
@@ -34,6 +34,10 @@
       "types": "./_lib/jsx/jsx-dev-runtime/index.d.ts",
       "default": "./_lib/jsx/jsx-dev-runtime/index.js"
     },
+    "./next": {
+      "types": "./_lib/next/index.d.ts",
+      "default": "./_lib/next/index.js"
+    },
     "./vercel": {
       "types": "./_lib/vercel/index.d.ts",
       "default": "./_lib/vercel/index.js"
