+++ templates/next/app/page.tsx
@@ -1,5 +1,17 @@
+import { getFrameMetadata } from 'frog/next'
+import type { Metadata } from 'next'
 import Image from 'next/image'
+
 import styles from './page.module.css'
+
+export async function generateMetadata(): Promise<Metadata> {
+  const frameTags = await getFrameMetadata(
+    `${process.env.VERCEL_URL || 'http://localhost:3000'}/api`,
+  )
+  return {
+    other: frameTags,
+  }
+}
 
 export default function Home() {
   return (
